Rate-limit compactions.
With a large compaction queue, kicking off a compaction of all the rows
in parallel is adversely impacting the TSD and HBase.  The sudden surge
of requests (get/put/delete) tends to cause lots of regions to split at
roughly the same time, which isn't good.   The large number of requests
in flight puts unnecessary memory pressure on the TSD.

So right now the rate-limit is hardcoded (compile-time constant) and set
to 1000 rows per 10s interval (so around 100 rows/s).  This threshold is
completely arbitrary and ultimately needs to be changed dynamically with
the queue size, in order to ensure that the queue is emptied fast enough
to guarantee it doesn't grow unbounded for the current workload.

Change-Id: I3e8ce6b60fd8ec970ec863fa33c59e02e7428e09