Properly handle throttling requests from asynchbase.
Compactions rewrite so much data that they're very likely to cause
region splits in HBase.  When they do, they keep sending tons of
requests to the HBase client, which eventually makes them all fail
with a PleaseThrottleException, due to the nature of its completely
non-blocking asynchronous implementation.  When we get such this
exception, we simply re-queue the row for compaction again, to try
to compact it later.

Change-Id: Ie7f5c4291c82ce9485243b7265763a399150f3d1